requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
bleach==2.0.0
chardet==2.3.0
Django==1.10.6
django-bootstrap-form==3.2.1
django-contrib-comments==1.8.0
django-crispy-forms==1.6.1
filebrowser-safe==0.4.7
future==0.16.0
grappelli-safe==0.4.6
html5lib==0.999999999
lxml==3.6.0
Mezzanine==4.2.3
numpy==1.12.1
oauthlib==2.0.2
olefile==0.44
packaging==16.8
pandas==0.19.2
Pillow==4.0.0
pyparsing==2.2.0
python-dateutil==2.6.0
python-pptx==0.6.5
pytz==2016.10
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
South==1.0.2
tzlocal==1.3
webencodings==0.5
xlrd==1.0.0
XlsxWriter==0.9.6
